feat(search): 라우트 변경 /profile/search -> /search
