+++ app/features/search/components/ProfileCard.module.scss
