+++ app/features/search/pages/search.module.scss
