+++ app/routes/search.tsx
@@ -1,5 +1,5 @@
-import { searchLoader } from 'app/features/profile/loader';
-import Search from 'app/features/profile/pages/search';
+import { searchLoader } from 'app/features/search/loader';
+import Search from 'app/features/search/pages/search';
 
 export const meta = () => [{ title: '유저 검색' }];
